created orders endpoints & procedure

+++ database/tables/cart.sql
@@ -12,7 +12,7 @@ CREATE TABLE dbo.cart
 (
  id VARCHAR(50) NOT NULL,
  user_id VARCHAR(50) NOT NULL,
- total FLOAT NOT NULL,
+ total INT NOT NULL,
  created_at TIMESTAMP NOT NULL,
  
  PRIMARY KEY (id),
